// app/Interfaces/RoomDetails.ts
export interface RoomDetails {
    status: string;
    _rev: string;
    _id: string;
    id: string;
    type: string;
    pricePerNight: number;
    mainImage: { _type: string; asset: { _ref: string; _type: string } };
    _createdAt: string;
    _type: string;
    _updatedAt: string;
    confort: string;
  }